backend/requirements.txt
fastapi
uvicorn
sqlalchemy
python-dotenv
pydantic
psycopg2-binary
duckduckgo-search
